backend/pyproject.toml
[tool.poetry]
name = "cadenza"
version = "0.1.0"
description = ""
authors = ["Reinhart Ulrich"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
beautifulsoup4 = ">=4.12.2"
colorama = ">=0.4.6"
duckduckgo_search = ">=4.1.1"
md2pdf = ">=1.0.1"
openai = ">=1.3.3"
playwright = ">=1.39.0"
python-dotenv = ">=1.0.0"
pyyaml = ">=6.0.1"
uvicorn = ">=0.24.0.post1"
pydantic = ">=2.5.1"
fastapi = ">=0.104.1"
python-multipart = ">=0.0.6"
markdown = ">=3.5.1"
langchain = "^0.2"
langgraph = ">=0.0.29,<0.1"
langchain_community = "^0.2"
langchain-openai = "^0.1"
langchain-google-genai = "^1"
tavily-python = "^0.3.5"
permchain = ">=0.0.6"
arxiv = ">=2.0.0"
PyMuPDF = ">=1.23.6"
requests = ">=2.31.0"
jinja2 = ">=3.1.2"
aiofiles = ">=23.2.1"
newspaper3k = ">=0.2.8"
SQLAlchemy = ">=2.0.28"
mistune = "^3.0.2"
htmldocx = "^0.0.6"
python-docx = "^1.1.0"
lxml = { version = ">=4.9.2", extras = ["html_clean"] }
unstructured = "^0.13.0"
tiktoken = ">=0.7.0"
asyncpg = "^0.29.0"
sqlmodel = "^0.0.19"
pydantic-settings = "^2.3.3"
json-repair = "^0.25.1"
langchain-groq = "^0.1.5"
langchain-together = "^0.1.3"
langchain-huggingface = "^0.0.3"
langchain-mistralai = "^0.1.8"
langchain-anthropic = "^0.1.15"
langchain-google-vertexai = "^1.0.5"
langchain-cohere = "^0.1.8"
langchain-fireworks = "^0.1.3"
langchain-aws = "^0.1.7"
langchainhub = "^0.1.20"
alembic = "^1.13.1"
gunicorn = "^22.0.0"
azure-search = "^1.0.0b2"
azure-search-documents = "^11.4.0"
aiohttp = "^3.9.5"
langfuse = "^2.39.0"
highcharts-core = "^1.9.2"
highchartexport = "^0.0.5"
pypdf = "^4.3.1"
faiss-cpu = "^1.8.0.post1"
rapidocr-onnxruntime = "^1.3.24"
exa-py = "^1.0.16"
docx2txt = "^0.8"
python-pptx = "^0.6.23"
pyjwt = "^2.9.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.10"
pytest = "^8.2.2"
httpx = "^0.27.0"
pytest-cov = "^5.0.0"
aiosqlite = "^0.20.0"
pytest-anyio = "^0.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
